move CacheWarningCallStack into its own file
