+++ lib/streamlit/runtime/caching/cache_data_api.py
@@ -42,13 +42,13 @@ from streamlit.runtime.caching.cache_utils import (
     CachedFunction,
     CachedResult,
     CacheMessagesCallStack,
-    CacheWarningCallStack,
     ElementMsgData,
     MsgData,
     MultiCacheResults,
     create_cache_wrapper,
     ttl_to_seconds,
 )
+from streamlit.runtime.caching.cache_warning_call_stack import CacheWarningCallStack
 from streamlit.runtime.metrics_util import gather_metrics
 from streamlit.runtime.scriptrunner.script_run_context import get_script_run_ctx
 from streamlit.runtime.stats import CacheStat, CacheStatsProvider
