+++ lib/streamlit/runtime/caching/cache_utils.py
@@ -33,12 +33,10 @@ from typing_extensions import Protocol, runtime_checkable
 
 import streamlit as st
 from streamlit import runtime, type_util, util
-from streamlit.elements import NONWIDGET_ELEMENTS, WIDGETS
 from streamlit.elements.spinner import spinner
 from streamlit.logger import get_logger
 from streamlit.proto.Block_pb2 import Block
 from streamlit.runtime.caching.cache_errors import (
-    CachedStFunctionWarning,
     CacheError,
     CacheKeyNotFoundError,
     CacheReplayClosureError,
@@ -49,6 +47,7 @@ from streamlit.runtime.caching.cache_errors import (
     get_cached_func_name_md,
 )
 from streamlit.runtime.caching.cache_type import CacheType
+from streamlit.runtime.caching.cache_warning_call_stack import CacheWarningCallStack
 from streamlit.runtime.caching.hashing import update_hash
 from streamlit.runtime.scriptrunner.script_run_context import (
     ScriptRunContext,
@@ -435,113 +434,6 @@ def create_cache_wrapper(cached_func: CachedFunction) -> Callable[..., Any]:
     wrapper.clear = clear  # type: ignore
 
     return wrapper
-
-
-class CacheWarningCallStack(threading.local):
-    """A utility for warning users when they call `st` commands inside
-    a cached function. Internally, this is just a counter that's incremented
-    when we enter a cache function, and decremented when we exit.
-
-    Data is stored in a thread-local object, so it's safe to use an instance
-    of this class across multiple threads.
-    """
-
-    def __init__(self, cache_type: CacheType):
-        self._cached_func_stack: list[types.FunctionType] = []
-        self._suppress_st_function_warning = 0
-        self._cache_type = cache_type
-        self._allow_widgets: int = 0
-
-    def __repr__(self) -> str:
-        return util.repr_(self)
-
-    @contextlib.contextmanager
-    def calling_cached_function(self, func: types.FunctionType) -> Iterator[None]:
-        self._cached_func_stack.append(func)
-        try:
-            yield
-        finally:
-            self._cached_func_stack.pop()
-
-    @contextlib.contextmanager
-    def suppress_cached_st_function_warning(self) -> Iterator[None]:
-        self._suppress_st_function_warning += 1
-        try:
-            yield
-        finally:
-            self._suppress_st_function_warning -= 1
-            assert self._suppress_st_function_warning >= 0
-
-    @contextlib.contextmanager
-    def maybe_suppress_cached_st_function_warning(
-        self, suppress: bool
-    ) -> Iterator[None]:
-        if suppress:
-            with self.suppress_cached_st_function_warning():
-                yield
-        else:
-            yield
-
-    def maybe_show_cached_st_function_warning(
-        self,
-        dg: "st.delta_generator.DeltaGenerator",
-        st_func_name: str,
-    ) -> None:
-        """If appropriate, warn about calling st.foo inside @memo.
-
-        DeltaGenerator's @_with_element and @_widget wrappers use this to warn
-        the user when they're calling st.foo() from within a function that is
-        wrapped in @st.cache.
-
-        Parameters
-        ----------
-        dg : DeltaGenerator
-            The DeltaGenerator to publish the warning to.
-
-        st_func_name : str
-            The name of the Streamlit function that was called.
-
-        """
-        # There are some elements not in either list, which we still want to warn about.
-        # Ideally we will fix this by either updating the lists or creating a better
-        # way of categorizing elements.
-        if st_func_name in NONWIDGET_ELEMENTS:
-            return
-        if st_func_name in WIDGETS and self._allow_widgets > 0:
-            return
-
-        if len(self._cached_func_stack) > 0 and self._suppress_st_function_warning <= 0:
-            cached_func = self._cached_func_stack[-1]
-            self._show_cached_st_function_warning(dg, st_func_name, cached_func)
-
-    def _show_cached_st_function_warning(
-        self,
-        dg: "st.delta_generator.DeltaGenerator",
-        st_func_name: str,
-        cached_func: types.FunctionType,
-    ) -> None:
-        # Avoid infinite recursion by suppressing additional cached
-        # function warnings from within the cached function warning.
-        with self.suppress_cached_st_function_warning():
-            e = CachedStFunctionWarning(self._cache_type, st_func_name, cached_func)
-            dg.exception(e)
-
-    @contextlib.contextmanager
-    def allow_widgets(self) -> Iterator[None]:
-        self._allow_widgets += 1
-        try:
-            yield
-        finally:
-            self._allow_widgets -= 1
-            assert self._allow_widgets >= 0
-
-    @contextlib.contextmanager
-    def maybe_allow_widgets(self, allow: bool) -> Iterator[None]:
-        if allow:
-            with self.allow_widgets():
-                yield
-        else:
-            yield
 
 
 """
